feat: enhance navigation accessibility and resolve CI/lint issues
UX Improvements:
- Added "Skip to content" link for keyboard users.
- Replaced Link with NavLink for better ARIA support.
- Added visual cues (borders) for active nav states.
- Improved focus indicators for all navigation links.

CI & Maintenance:
- Fixed CodeQL injection vulnerability in GitFlow action.
- Resolved pre-existing lint errors and warnings across multiple files.
- Synchronized package-lock.json with package.json.
- Configured Snyk scan to continue on error.
- Moved useToast to a separate hook to fix Fast Refresh warnings.
