+++ src/hooks/useToast.ts
@@ -1,0 +1,22 @@
+import { useState } from 'react'
+
+export interface Toast {
+  id: string
+  message: string
+  type: 'success' | 'error' | 'info'
+}
+
+export const useToast = () => {
+  const [, setToasts] = useState<Toast[]>([])
+
+  const addToast = (message: string, type: Toast['type'] = 'info') => {
+    const id = Math.random().toString(36).substr(2, 9)
+    setToasts((prev) => [...prev, { id, message, type }])
+  }
+
+  const removeToast = (id: string) => {
+    setToasts((prev) => prev.filter((toast) => toast.id !== id))
+  }
+
+  return { addToast, removeToast }
+}
